refactor: sqlite.ts 拆出成本账本 cost-ledger.ts（token_cost 表族：建表迁移+写入清理+四路聚合）——MemoryDb 四方法一行委托，接口与调用方零改动（体检 P1 第一刀）
